add selection event for theme + article
