+++ packages/code-du-travail-frontend/src/search/SearchResults/Law.js
@@ -1,6 +1,5 @@
 import React from "react";
 import Link from "next/link";
-import { getRouteBySource } from "@cdt/sources";
 import {
   CardList,
   Container,
@@ -8,37 +7,45 @@ import {
   Tile,
   Wrapper
 } from "@socialgouv/react-ui";
+import { summarize, reportSelectionToMatomo } from "../utils";
+import { getRouteBySource } from "@cdt/sources";
 
-import { summarize } from "../utils";
-
-export const Law = ({ items, query }) => (
-  <Section>
-    <Container>
-      <Wrapper>
-        <CardList
-          leftStripped
-          title="Que dit le code du travail&nbsp;?"
-          columns={2}
-        >
-          {items.map(({ slug, title, source, description }) => (
-            <Link
-              key={slug}
-              href={{
-                pathname: `${getRouteBySource(source)}/[slug]`,
-                query: query ? { q: query } : null
-              }}
-              as={`/${getRouteBySource(source)}/${slug}${
-                query ? `?q=${query}` : ""
-              }`}
-              passHref
-            >
-              <Tile wide title={title}>
-                {summarize(description)}
-              </Tile>
-            </Link>
-          ))}
-        </CardList>
-      </Wrapper>
-    </Container>
-  </Section>
-);
+export const Law = ({ items, query }) => {
+  return (
+    <Section>
+      <Container>
+        <Wrapper>
+          <CardList
+            leftStripped
+            title="Que dit le code du travail&nbsp;?"
+            columns={2}
+          >
+            {items.map(({ slug, title, source, description, url, algo }) => (
+              <Link
+                key={slug}
+                href={{
+                  pathname: `${getRouteBySource(source)}/[slug]`,
+                  query: query ? { q: query } : null
+                }}
+                as={`/${getRouteBySource(source)}/${slug}${
+                  query ? `?q=${query}` : ""
+                }`}
+                passHref
+              >
+                <Tile
+                  wide
+                  title={title}
+                  onClick={() =>
+                    reportSelectionToMatomo(source, slug, url, algo)
+                  }
+                >
+                  {summarize(description)}
+                </Tile>
+              </Link>
+            ))}
+          </CardList>
+        </Wrapper>
+      </Container>
+    </Section>
+  );
+};
