+++ packages/code-du-travail-frontend/src/search/SearchResults/index.js
@@ -5,6 +5,7 @@ import { Results } from "./Results";
 import { Law } from "./Law";
 import { Themes } from "./Themes";
 import { matopush } from "../../piwik";
+import { SOURCES, getRouteBySource } from "@cdt/sources";
 
 const SearchResults = ({
   items: { documents, themes, articles },
@@ -13,9 +14,13 @@ const SearchResults = ({
 }) => {
   const router = useRouter();
   useEffect(() => {
-    const toLogCandidate = ({ source, slug, algo }) => {
+    const toLogCandidate = ({ url, source, slug, algo }) => {
+      const trackedUrl =
+        source === SOURCES.EXTERNALS
+          ? url
+          : `/${getRouteBySource(source)}/${slug}`;
       return {
-        slug: `${source}/${slug}`,
+        slug: trackedUrl,
         algo
       };
     };
